libre: Extract char class/type lookup into dialect-specific file.
The CHAR_TYPE (distict from CHAR_CLASS) stuff will probably get
absorbed into CHAR_CLASS in the near future, but this is a step
along the way.

Also:

- Fix the prototype for `re_ast_print`, which got renamed.

- Add a missing `#include <ctype.h>` and `<string.h>`.

- Mark `<err-unsupported>` as unused in the literal dialect parser.

- Fix an error handling path in the SQL dialect parser.

- Add an assertion for a NULL AST (which can be triggered by
  improper error handling in the parser).
