+++ src/libre/dialect/glob/re_dialect_glob.c
@@ -1,0 +1,17 @@
+#include "../../re_char_class.h"
+
+int
+re_char_class_glob(const char *name, enum ast_char_class_id *id)
+{
+	(void)name;
+	(void)id;
+	return 0;
+}
+
+int
+re_char_type_glob(const char *name, enum ast_char_type_id *id)
+{
+	(void)name;
+	(void)id;
+	return 0;
+}
