+++ src/libre/dialect/literal/parser.sid
@@ -112,7 +112,7 @@
 	!<err-expected-closeflags>;
 	!<err-expected-groupbody>;
 	<err-expected-eof>;
-	<err-unsupported>;
+	!<err-unsupported>;
 
 	!<mark-group>: (:pos, :pos) -> ();
 	!<mark-range>: (:pos, :pos) -> ();
