+++ src/libre/dialect/literal/re_dialect_literal.c
@@ -1,0 +1,17 @@
+#include "../../re_char_class.h"
+
+int
+re_char_class_literal(const char *name, enum ast_char_class_id *id)
+{
+	(void)name;
+	(void)id;
+	return 0;
+}
+
+int
+re_char_type_literal(const char *name, enum ast_char_type_id *id)
+{
+	(void)name;
+	(void)id;
+	return 0;
+}
