+++ src/libre/parser.act
@@ -54,6 +54,8 @@
 	#define LX_INIT    CAT(LX_PREFIX, _init)
 
 	#define DIALECT_PARSE  CAT(parse_re_, DIALECT)
+	#define DIALECT_CHAR_CLASS  CAT(re_char_class_, DIALECT)
+	#define DIALECT_CHAR_TYPE   CAT(re_char_type_, DIALECT)
 
 	/* XXX: get rid of this; use same %entry% for all grammars */
 	#define DIALECT_ENTRY CAT(p_re__, DIALECT)
@@ -684,8 +686,7 @@
 	@};
 
 	CHAR_TYPE: () -> (id :ast_char_type_id, start :pos, end :pos) = @{
-		/* enum ast_char_class_type_id type_id; */
-		if (!re_char_class_type_id_of_name(lex_state->buf.a, &@id)) {
+		if (!DIALECT_CHAR_TYPE(lex_state->buf.a, &@id)) {
 			@!;
 		}
 
@@ -694,8 +695,7 @@
 	@};
 
 	NAMED_CHAR_CLASS: () -> (id :ast_char_class_id, start :pos, end :pos) = @{
-		/* enum ast_char_class_type_id type_id; */
-		if (!re_char_class_id_of_name(lex_state->buf.a, &@id)) {
+		if (!DIALECT_CHAR_CLASS(lex_state->buf.a, &@id)) {
 			@!;
 		}
